fix: rearrange file system
