+++ Assets/Scripts/Players/Interface.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e793b3834bab1c54c861c2c5618a3928
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
