+++ Assets/Scripts/Players/Interface/BasePlayerController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9cca0ffd3ea23b24ba9a7a4a62df5ee0
+guid: d096c2edb51d8ee47b6eb4007b5d3888
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
