+++ Assets/Scripts/Players/Player.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cb93178e5caefc04599c6460a5245e07
+guid: e1c06956d10b31444984f86f7c355e19
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
