+++ Assets/Scripts/Players/PlayerController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3efc27fbab83d3648b3a4205f902d874
+guid: ec4df0e9642fda24e9eab5e4e747073f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
