+++ Assets/Scripts/Spells.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 277cf5e99faaf384ba97769f89bae46c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
