+++ Assets/Scripts/Spells/Flux.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cf085aba8cecd3048ab0ade6f6e9181f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
