+++ Assets/Scripts/Spells/Flux/FluxShot.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0936fa5babff5e84988bae9309cae705
+guid: cf4df7e20e03d8e4f94fd400b35ae091
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
