+++ Assets/Scripts/Spells/Flux/FluxShotBullet.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8f04eb23e3a71dd4fa42338714e99671
+guid: 0e84af4441d8f034fb074282d1904b99
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
