+++ Assets/Scripts/Spells/Interface.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 393b6b230aa02ec42b6e432dd42c644a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
