+++ Assets/Scripts/Spells/Interface/BaseProjectileSpell.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 41390dd645cf0e64eab22b0294256450
+guid: b573d745d94c50240a83ee3b28223fea
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
