+++ Assets/Scripts/Spells/ShootingSkill.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public abstract class ShootingSkill : MonoBehaviour
 {
     public abstract void Shoot();
-    
+
     protected virtual BaseProjectileSpell SpawnProjectile(BaseProjectileSpell prefab, Vector3 position)
     {
         var spell = Instantiate(prefab, position, Quaternion.identity);
