+++ Assets/Scripts/Spells/ShootingSkill.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 76a476f1f3d2f2c44b6c74c303169ef9
+guid: b547fc7ea79c2ab4382127e1a833939f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
