+++ Assets/Scripts/Utils/ForceCameraRatio.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3e75455201130cd43864c935ba61835b
+guid: 404e331b08bb3a240a347b7c7fb90f46
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
