+++ Assets/Scripts/Utils/IDamageable.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 84d4afd561852c74f8383ca7651b8f4b
+guid: 22b0f09898eb93542a75a80ebe0f2dd4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
